chore(refactor): Seperated directories
